requirements.txt
Python==3.8.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-widget-tweaks==1.5.0
idna==3.6
oauthlib==3.2.2
phonenumbers==8.13.32
pillow==10.2.0
pip==24.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-slugify==8.0.4
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
setuptools==69.1.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
urllib3==2.2.1
wheel==0.42.0
